Page design
Code is a bit messy, and I will clean it up later (tonight/tomorrow). Work within the section header of each pages .html, unless you are implementing forms. When you do so, look at the .html/.css of the donate page to see an example. Note that we are not implementing any functionality at this point. Let me know if there are any issues with the navigation bar or the responsive triggers.